chore: sentry config update
